+++ js/consts.js
@@ -1,0 +1,44 @@
+const COUNT_POSTS = 25;
+
+const AvatarId = {
+  MIN: 1,
+  MAX: 6,
+};
+
+const MessageCount = {
+  MIN: 1,
+  MAX: 2,
+};
+
+const CommentsCount = {
+  MIN: 0,
+  MAX: 30,
+};
+
+const LikesCount = {
+  MIN: 15,
+  MAX: 200,
+};
+
+const MESSAGES = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Анна',
+  'Виктор',
+  'Александр',
+  'Наталья',
+  'Артемий',
+  'Евангелина',
+  'Владимир',
+  'Анастасия',
+  'Егор',
+  'Дарья'
+];
+
+export {COUNT_POSTS, AvatarId, MessageCount, CommentsCount, LikesCount, MESSAGES, NAMES};
